refactor(repositories/essentia): join essentia persistence related functions on a single file
Assure data directory exists
